requirements.txt
psycopg2-binary
pandas
fastapi
uvicorn
google-generativeai
